refactor: common workspace menu & memo menu
