+++ web/src/beta/features/Navbar/LeftSection/Profile/index.tsx
@@ -1,7 +1,8 @@
+import useWorkspaceManagementMenu from "@reearth/beta/hooks/useWorkspaceManagementMenu";
 import { PopupMenu, PopupMenuItem } from "@reearth/beta/lib/reearth-ui";
 import { config } from "@reearth/services/config";
 import { useT } from "@reearth/services/i18n";
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 import { useNavigate } from "react-router-dom";
 
 import { Workspace } from "../../types";
@@ -36,55 +37,56 @@ const HeaderProfile: React.FC<Props> = ({
     }
   }, [currentWorkspace?.id, navigate]);
 
-  const workspaceManagementMenu: PopupMenuItem[] = [
-    {
-      id: "workspaceSettings",
-      title: t("Workspace Settings"),
-      icon: "setting",
-      hasBorderBottom: true,
-      onClick: () => navigate(`/settings/workspaces/${currentWorkspace?.id}`)
-    },
-    {
-      id: "accountSettings",
-      title: t("Account Settings"),
-      icon: "user",
-      onClick: () => navigate("/settings/account")
-    }
-  ];
+  const { workspaceManagementMenu } = useWorkspaceManagementMenu({
+    workspaceId: currentWorkspace?.id
+  });
 
-  const popupMenu: PopupMenuItem[] = [
-    {
-      icon: "arrowLeftRight",
-      id: "switchWorkspace",
-      subItem: workspaces.map((w) => {
-        return {
-          customSubMenuLabel: w.personal ? t("Personal") : t("Team Workspace"),
-          customSubMenuOrder: w.personal ? 0 : 1,
-          id: w.id as string,
-          title: w.name ?? t("Unknown"),
-          hasCustomSubMenu: true,
-          personal: w.personal,
-          selected: currentWorkspace?.id === w.id,
-          onClick: () => w.id && handleWorkspaceChange(w.id)
-        };
-      }),
-      title: t("Switch workspace")
-    },
-    ...(config()?.disableWorkspaceManagement ? [] : workspaceManagementMenu),
-    {
-      icon: "exit",
-      id: "logOut",
-      hasBorderBottom: true,
-      onClick: onSignOut,
-      title: t("Log out")
-    },
-    {
-      icon: "file",
-      id: "assetManagement",
-      onClick: handleAssetManager,
-      title: t("Asset management")
-    }
-  ];
+  const popupMenu: PopupMenuItem[] = useMemo(
+    () => [
+      {
+        icon: "arrowLeftRight",
+        id: "switchWorkspace",
+        subItem: workspaces.map((w) => {
+          return {
+            customSubMenuLabel: w.personal
+              ? t("Personal")
+              : t("Team Workspace"),
+            customSubMenuOrder: w.personal ? 0 : 1,
+            id: w.id as string,
+            title: w.name ?? t("Unknown"),
+            hasCustomSubMenu: true,
+            personal: w.personal,
+            selected: currentWorkspace?.id === w.id,
+            onClick: () => w.id && handleWorkspaceChange(w.id)
+          };
+        }),
+        title: t("Switch workspace")
+      },
+      ...(config()?.disableWorkspaceManagement ? [] : workspaceManagementMenu),
+      {
+        icon: "exit",
+        id: "logOut",
+        hasBorderBottom: true,
+        onClick: onSignOut,
+        title: t("Log out")
+      },
+      {
+        icon: "file",
+        id: "assetManagement",
+        onClick: handleAssetManager,
+        title: t("Asset management")
+      }
+    ],
+    [
+      currentWorkspace?.id,
+      handleAssetManager,
+      handleWorkspaceChange,
+      onSignOut,
+      t,
+      workspaces,
+      workspaceManagementMenu
+    ]
+  );
 
   return <PopupMenu label={currentWorkspace?.name} menu={popupMenu} />;
 };
