+++ web/src/beta/pages/WorkspaceSettingPage/index.tsx
@@ -1,7 +1,7 @@
 import NotFound from "@reearth/beta/features/NotFound";
 import WorkspaceSetting from "@reearth/beta/features/WorkspaceSetting";
 import { config } from "@reearth/services/config";
-import { FC } from "react";
+import { FC, useMemo } from "react";
 import { useParams } from "react-router-dom";
 
 import Page from "../Page";
@@ -12,11 +12,12 @@ type WorkspaceSettingPageProps = {
 
 const WorkspaceSettingPage: FC<WorkspaceSettingPageProps> = ({ tab }) => {
   const { workspaceId } = useParams<{ workspaceId: string }>();
+  const disabled = useMemo(() => config()?.disableWorkspaceManagement, []);
   return (
     <Page
       workspaceId={workspaceId}
       renderItem={({ workspaceId }) =>
-        config()?.disableWorkspaceManagement ? (
+        disabled ? (
           <NotFound />
         ) : (
           <WorkspaceSetting tab={tab} workspaceId={workspaceId} />
